attempt to fix conflicts with mirror URL callbacks

+++ app/models/project.rb
@@ -90,7 +90,8 @@ class Project < ActiveRecord::Base
   after_destroy :remove_pages
 
   after_update :update_forks_visibility_level
-  after_update :remove_mirror_repository_reference, if: :import_url_changed?
+  after_update :remove_mirror_repository_reference,
+               if: ->(project) { project.mirror? && project.import_url_changed? }
 
   ActsAsTaggableOn.strict_case_match = true
   acts_as_taggable_on :tags
@@ -1261,6 +1262,8 @@ class Project < ActiveRecord::Base
   end
 
   def remove_mirror_repository_reference
+    # return if it's the first assignment
+    return unless changes['import_url'] && changes['import_url'].first
     repository.remove_remote(Repository::MIRROR_REMOTE)
   end
